requirements.txt
# Azure SDK Dependencies
azure-functions>=1.18.0
azure-eventhub>=5.11.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
